fixed64: codemod pass 8 -- cascade Fixed->scalar narrowings
21 more Fixed->scalar casts revealed after the division/multiply fixes
(single + multiline RHS ranges). Behavior-preserving; flag-off green;
tests green.

+++ src/zandronum/src/g_doom/a_painelemental.cpp
@@ -90,8 +90,8 @@ void A_PainShootSkull (AActor *self, angle_t angle, const PClass *spawntype, int
 	// okay, there's room for another one
 	an = angle >> ANGLETOFINESHIFT;
 	
-	prestep = (int)(4*FRACUNIT +
-		3*(self->radius + GetDefaultByType(spawntype)->radius)/2);
+	prestep = (int)((int)(4*FRACUNIT +
+		3*(self->radius + GetDefaultByType(spawntype)->radius)/2));
 	
 	x = self->x + FixedMul (prestep, finecosine[an]);
 	y = self->y + FixedMul (prestep, finesine[an]);
