+++ src/zandronum/src/p_enemy.cpp
@@ -3307,8 +3307,8 @@ DEFINE_ACTION_FUNCTION(AActor, A_MonsterRail)
 	{
 		// We probably won't hit the target, but aim at it anyway so we don't look stupid.
 		FVector2 xydiff((double)(self->target->x - self->x),(double)( self->target->y - self->y));
-		double zdiff = (double)((self->target->z + (self->target->height>>1)) -
-						(self->z + (self->height>>1) - self->floorclip));
+		double zdiff = (double)((double)((self->target->z + (self->target->height>>1)) -
+						(self->z + (self->height>>1) - self->floorclip)));
 		self->pitch = int(atan2(zdiff, xydiff.Length()) * ANGLE_180 / -M_PI);
 	}
